+++ core/src/algorithms/poll_math.rs
@@ -9,21 +9,21 @@ const MIN_EFFECTIVE_DT_MS: u64 = 500;
 const QUANTIZATION_STEP_MS: u64 = 100;
 const HYSTERESIS_THRESHOLD_MS: u64 = 200;
 const JITTER_PERCENT: u64 = 5;
-const ATTACK_COEFF: f64 = 1.0;
-const DECAY_COEFF: f64 = 0.1;
+const ATTACK_COEFF: f32 = 1.0;
+const DECAY_COEFF: f32 = 0.1;
 
 pub struct AdaptivePoller {
     current_interval: u64,
-    last_pressure: f64,
+    last_pressure: f32,
     last_tick: Instant,
     target_interval: u64,
-    weight_pressure: f64,
-    weight_derivative: f64,
+    weight_pressure: f32,
+    weight_derivative: f32,
     rng_state: u64,
 }
 
 impl AdaptivePoller {
-    pub fn new(weight_pressure: f64, weight_derivative: f64) -> Self {
+    pub fn new(weight_pressure: f32, weight_derivative: f32) -> Self {
         let start_seed = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .unwrap_or_default()
@@ -49,7 +49,7 @@ impl AdaptivePoller {
         let limit = range * 2;
         self.rng_state % (limit + 1)
     }
-    pub fn calculate_next_interval(&mut self, current_pressure: f64, avg300: f64) -> u64 {
+    pub fn calculate_next_interval(&mut self, current_pressure: f32, avg300: f32) -> u64 {
         let now = Instant::now();
         let elapsed_ms = now.duration_since(self.last_tick).as_millis() as u64;
         if elapsed_ms > (self.current_interval + SLEEP_TOLERANCE_MS) {
@@ -67,19 +67,19 @@ impl AdaptivePoller {
             (MIN_POLLING_MS, MAX_POLLING_MS)
         };
         let effective_dt_ms = elapsed_ms.max(MIN_EFFECTIVE_DT_MS);
-        let dt_sec = effective_dt_ms as f64 / 1000.0;
+        let dt_sec = effective_dt_ms as f32 / 1000.0;
         let rate_change = (current_pressure - self.last_pressure) / dt_sec;
         let prediction = current_pressure + (rate_change * ATTACK_COEFF);
         let p_term = prediction * self.weight_pressure;
         let d_term = rate_change.abs() * self.weight_derivative;
         let priority_score = (p_term + d_term).clamp(0.0, 100.0);
         let raw_interval =
-            dynamic_max as f64 - ((priority_score / 100.0) * (dynamic_max - dynamic_min) as f64);
-        let target = if raw_interval < self.target_interval as f64 {
+            dynamic_max as f32 - ((priority_score / 100.0) * (dynamic_max - dynamic_min) as f32);
+        let target = if raw_interval < self.target_interval as f32 {
             raw_interval
         } else {
             (raw_interval * DECAY_COEFF)
-                + (self.target_interval as f64 * (ATTACK_COEFF - DECAY_COEFF))
+                + (self.target_interval as f32 * (ATTACK_COEFF - DECAY_COEFF))
         };
         self.target_interval = target as u64;
         let diff = (self.target_interval as i64 - self.current_interval as i64).abs();
@@ -92,8 +92,8 @@ impl AdaptivePoller {
         self.apply_discrete_math_mut(self.current_interval, dynamic_min, dynamic_max)
     }
     fn apply_discrete_math_mut(&mut self, interval: u64, min_limit: u64, max_limit: u64) -> u64 {
-        let quantized = ((interval as f64 / QUANTIZATION_STEP_MS as f64).round()
-            * QUANTIZATION_STEP_MS as f64) as u64;
+        let quantized = ((interval as f32 / QUANTIZATION_STEP_MS as f32).round()
+            * QUANTIZATION_STEP_MS as f32) as u64;
         let clamped = quantized.clamp(min_limit, max_limit);
         let jitter_range = (clamped * JITTER_PERCENT) / 100;
         let noise = self.next_random(jitter_range);
